requirements.txt
djangorestframework
django
django-debug-toolbar
